+++ src/ext/java/org/opentripplanner/ext/siri/SiriFuzzyTripMatcher.java
@@ -6,14 +6,16 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.ArrayList;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Route;
 import org.opentripplanner.model.Station;
+import org.opentripplanner.model.TransitMode;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripPattern;
+import org.opentripplanner.model.calendar.ServiceDate;
 import org.opentripplanner.routing.RoutingService;
 import org.opentripplanner.routing.algorithm.raptor.transit.mappers.DateMapper;
-import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.trippattern.TripTimes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -208,29 +210,27 @@ public class SiriFuzzyTripMatcher {
 
                 TripPattern tripPattern = index.getPatternForTrip().get(trip);
 
-                    String currentTripId = getUnpaddedTripId(trip.getId().getId());
+                String currentTripId = getUnpaddedTripId(trip.getId().getId());
 
-                    if (mappedTripsCache.containsKey(currentTripId)) {
-                        mappedTripsCache.get(currentTripId).add(trip);
-                    } else {
-                        Set<Trip> initialSet = new HashSet<>();
-                        initialSet.add(trip);
-                        mappedTripsCache.put(currentTripId, initialSet);
-                    }
+                if (mappedTripsCache.containsKey(currentTripId)) {
+                    mappedTripsCache.get(currentTripId).add(trip);
+                } else {
+                    Set<Trip> initialSet = new HashSet<>();
+                    initialSet.add(trip);
+                    mappedTripsCache.put(currentTripId, initialSet);
+                }
 
                 if (tripPattern != null &&
-                        (tripPattern.getMode().equals(TraverseMode.RAIL) /*||
-                                                    (trip.getTransportSubmode() != null &&
-                                                            trip.getTransportSubmode().equals(TransmodelTransportSubmode.RAIL_REPLACEMENT_BUS))*/)) {
-                    // TODO - SIRI: Add support for submode
-                    if (trip.getTripShortName() != null) {
-                        String tripShortName = trip.getTripShortName();
-                        if (mappedVehicleRefCache.containsKey(tripShortName)) {
-                            mappedVehicleRefCache.get(tripShortName).add(trip);
+                    tripPattern.matchesModeOrSubMode(TransitMode.RAIL, "railReplacementBus")
+                ) {
+                    if (trip.getInternalPlanningCode() != null) {
+                        String internalPlanningCode = trip.getInternalPlanningCode();
+                        if (mappedVehicleRefCache.containsKey(internalPlanningCode)) {
+                            mappedVehicleRefCache.get(internalPlanningCode).add(trip);
                         } else {
                             Set<Trip> initialSet = new HashSet<>();
                             initialSet.add(trip);
-                            mappedVehicleRefCache.put(tripShortName, initialSet);
+                            mappedVehicleRefCache.put(internalPlanningCode, initialSet);
                         }
                     }
                 }
@@ -380,4 +380,39 @@ public class SiriFuzzyTripMatcher {
         }
         return -1;
     }
+
+    /**
+     * Returns a match of tripIds that match the provided values.
+     */
+    public List<FeedScopedId> getTripIdForInternalPlanningCodeServiceDateAndMode(
+            String internalPlanningCode,
+            ServiceDate serviceDate,
+            TransitMode mode,
+            String transportSubmode
+    ) {
+        Set<Trip> cachedTripsBySiriId = getCachedTripsBySiriId(internalPlanningCode);
+
+        if (cachedTripsBySiriId.isEmpty()) {
+            cachedTripsBySiriId = getCachedTripsByVehicleRef(internalPlanningCode);
+        }
+
+        List<FeedScopedId> matches = new ArrayList<>();
+        for (Trip trip : cachedTripsBySiriId) {
+            final TripPattern tripPattern = routingService.getPatternForTrip().get(trip);
+            if (tripPattern.matchesModeOrSubMode(mode, transportSubmode)) {
+                Set<ServiceDate> serviceDates = routingService
+                        .getCalendarService()
+                        .getServiceDatesForServiceId(trip.getServiceId());
+                if (serviceDates.contains(serviceDate) &&
+                    trip.getInternalPlanningCode() != null &&
+                    trip.getInternalPlanningCode().equals(internalPlanningCode)) {
+                    matches.add(trip.getId());
+                }
+            }
+        }
+
+
+        return matches;
+    }
+
 }