+++ src/ext/java/org/opentripplanner/ext/siri/SiriTimetableSnapshotSource.java
@@ -16,6 +16,7 @@ import java.util.List;
 import java.util.Set;
 import java.util.TimeZone;
 import java.util.concurrent.locks.ReentrantLock;
+import org.opentripplanner.common.model.T2;
 import org.opentripplanner.model.Agency;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Operator;
@@ -29,12 +30,15 @@ import org.opentripplanner.model.TimetableSnapshotProvider;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.TripPattern;
 import org.opentripplanner.model.calendar.ServiceDate;
+import org.opentripplanner.model.TransitMode;
 import org.opentripplanner.routing.RoutingService;
 import org.opentripplanner.routing.algorithm.raptor.transit.mappers.DateMapper;
 import org.opentripplanner.routing.algorithm.raptor.transit.mappers.TransitLayerUpdater;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.trippattern.RealTimeState;
 import org.opentripplanner.routing.trippattern.TripTimes;
+import org.rutebanken.netex.model.BusSubmodeEnumeration;
+import org.rutebanken.netex.model.RailSubmodeEnumeration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.org.siri.siri20.ArrivalBoardingActivityEnumeration;
@@ -49,6 +53,23 @@ import uk.org.siri.siri20.VehicleActivityCancellationStructure;
 import uk.org.siri.siri20.VehicleActivityStructure;
 import uk.org.siri.siri20.VehicleModesEnumeration;
 import uk.org.siri.siri20.VehicleMonitoringDeliveryStructure;
+
+import java.time.ZonedDateTime;
+import java.util.ArrayList;
+import java.util.BitSet;
+import java.util.Date;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
+import java.util.TimeZone;
+import java.util.concurrent.locks.ReentrantLock;
+
+import static org.opentripplanner.ext.siri.SiriTransportModeMapper.mapTransitMainMode;
+import static org.opentripplanner.ext.siri.TimetableHelper.createModifiedStopTimes;
+import static org.opentripplanner.ext.siri.TimetableHelper.createModifiedStops;
+import static org.opentripplanner.ext.siri.TimetableHelper.createUpdatedTripTimes;
+import static org.opentripplanner.model.PickDrop.NONE;
+import static org.opentripplanner.model.PickDrop.SCHEDULED;
 
 /**
  * This class should be used to create snapshots of lookup tables of realtime data. This is
@@ -463,7 +484,12 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
 
         if (route == null) { // Route is unknown - create new
             route = new Route(routeId);
-            route.setType(getRouteType(estimatedVehicleJourney.getVehicleModes()));
+            T2<TransitMode, String> transitMode = getTransitMode(
+                estimatedVehicleJourney.getVehicleModes(),
+                replacedRoute
+            );
+            route.setMode(transitMode.first);
+            route.setNetexSubmode(transitMode.second);
             route.setOperator(operator);
 
             // TODO - SIRI: Is there a better way to find authority/Agency?
@@ -484,19 +510,6 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
 
         Trip trip = new Trip(tripId);
         trip.setRoute(route);
-
-        // TODO - SIRI: Set transport-submode based on replaced- and replacement-route
-//        if (replacedRoute != null) {
-//            if (replacedRoute.getType() >= 100 && replacedRoute.getType() < 200) { // Replaced-route is RAIL
-//                if (route.getType() == 100) {
-//                    // Replacement-route is also RAIL
-//                    trip.setTransportSubmode(TransmodelTransportSubmode.REPLACEMENT_RAIL_SERVICE);
-//                } else if (route.getType() == 700) {
-//                    // Replacement-route is BUS
-//                    trip.setTransportSubmode(TransmodelTransportSubmode.RAIL_REPLACEMENT_BUS);
-//                }
-//            }
-//        }
 
         ServiceDate serviceDate = getServiceDateForEstimatedVehicleJourney(estimatedVehicleJourney);
 
@@ -661,30 +674,43 @@ public class SiriTimetableSnapshotSource implements TimetableSnapshotProvider {
         return addTripToGraphAndBuffer(feedId, graph, trip, aimedStopTimes, addedStops, tripTimes, serviceDate);
     }
 
-    /*
-     * Resolves TransportMode from SIRI VehicleMode
+    /**
+     * Resolves TransitMode from SIRI VehicleMode
      */
-    private int getRouteType(List<VehicleModesEnumeration> vehicleModes) {
-        if (vehicleModes != null && !vehicleModes.isEmpty()) {
-            VehicleModesEnumeration vehicleModesEnumeration = vehicleModes.get(0);
-            switch (vehicleModesEnumeration) {
-                case RAIL:
-                    return 100;
-                case COACH:
-                    return 200;
-                case BUS:
-                    return 700;
-                case METRO:
-                    return 701;
-                case TRAM:
-                    return 900;
-                case FERRY:
-                    return 1000;
-                case AIR:
-                    return 1100;
+    private T2<TransitMode, String> getTransitMode(
+        List<VehicleModesEnumeration> vehicleModes, Route replacedRoute
+    ) {
+        TransitMode transitMode = mapTransitMainMode(vehicleModes);
+
+        String transitSubMode = resolveTransitSubMode(transitMode, replacedRoute);
+
+        return new T2<>(transitMode, transitSubMode);
+    }
+
+    /**
+     * Resolves submode based on added trips's mode and replacedRoute's mode
+     * @param transitMode Mode of the added trip
+     * @param replacedRoute Route that is being replaced
+     * @return String-representation of submode
+     */
+    private String resolveTransitSubMode(TransitMode transitMode, Route replacedRoute) {
+        if (replacedRoute != null) {
+
+            TransitMode replacedRouteMode = replacedRoute.getMode();
+
+            if (replacedRouteMode != null &&
+                replacedRouteMode.equals(TransitMode.RAIL)) { // Replaced-route is RAIL
+
+                if (transitMode.equals(TransitMode.RAIL)) {
+                    // Replacement-route is also RAIL
+                    return RailSubmodeEnumeration.REPLACEMENT_RAIL_SERVICE.value();
+                } else if (transitMode.equals(TransitMode.BUS)) {
+                    // Replacement-route is BUS
+                    return BusSubmodeEnumeration.RAIL_REPLACEMENT_BUS.value();
+                }
             }
         }
-        return 700;
+        return null;
     }
 
     private boolean handleModifiedTrip(Graph graph, String feedId, EstimatedVehicleJourney estimatedVehicleJourney) {