+++ src/main/java/org/opentripplanner/model/Stop.java
@@ -38,6 +38,8 @@ public final class Stop extends StationElement implements StopLocation {
    */
   private final TransitMode vehicleType;
 
+  private final String netexSubmode;
+
   private HashSet<BoardingArea> boardingAreas;
 
   public Stop(
@@ -52,7 +54,8 @@ public final class Stop extends StationElement implements StopLocation {
       Collection<FareZone> fareZones,
       String url,
       TimeZone timeZone,
-      TransitMode vehicleType
+      TransitMode vehicleType,
+      String netexSubmode
   ) {
     super(id, name, code, description, coordinate, wheelchairBoarding, level);
     this.platformCode = platformCode;
@@ -60,6 +63,7 @@ public final class Stop extends StationElement implements StopLocation {
     this.url = url;
     this.timeZone = timeZone;
     this.vehicleType = vehicleType;
+    this.netexSubmode = netexSubmode;
   }
 
   /** @see #stopForTest(String, double, double, Station) */
@@ -78,6 +82,7 @@ public final class Stop extends StationElement implements StopLocation {
         idAndName,
         null,
         new WgsCoordinate(lat, lon),
+        null,
         null,
         null,
         null,
@@ -138,5 +143,10 @@ public final class Stop extends StationElement implements StopLocation {
   @Override
   public Geometry getGeometry() {
     return GeometryUtils.getGeometryFactory().createPoint(getCoordinate().asJtsCoordinate());
+  }
+
+  @Override
+  public String getVehicleSubmode() {
+    return netexSubmode;
   }
 }