+++ src/main/java/org/opentripplanner/model/TripPattern.java
@@ -112,6 +112,10 @@ public final class TripPattern extends TransitEntity implements Cloneable, Seria
         return route.getMode();
     }
 
+    public final String getNetexSubmode() {
+        return route.getNetexSubmode();
+    }
+
     public LineString getHopGeometry(int stopPosInPattern) {
         if (hopGeometries != null) {
             return CompactLineString.uncompactLineString(
@@ -582,6 +586,12 @@ public final class TripPattern extends TransitEntity implements Cloneable, Seria
         return tripPatterns.stream()
             .map(TransitEntity::getId)
             .allMatch(t -> t != null && seen.add(t));
+    }
+
+    public boolean matchesModeOrSubMode(TransitMode mode, String transportSubmode) {
+        return getMode().equals(mode) || (
+                getNetexSubmode() != null && getNetexSubmode().equals(transportSubmode)
+        );
     }
 
     public String toString () {