+++ src/main/java/org/opentripplanner/model/plan/PageCursor.java
@@ -1,0 +1,160 @@
+package org.opentripplanner.model.plan;
+
+import java.io.ByteArrayInputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
+import java.time.Duration;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.ZonedDateTime;
+import java.util.Base64;
+import javax.annotation.Nullable;
+import org.opentripplanner.model.base.ToStringBuilder;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * This class hold all the information needed to page to the next/previous page. It is
+ * serialized as base64 when passed on to the client. The base64 encoding is done to prevent the
+ * client from using the information inside the cursor.
+ * <p>
+ * This class is internal to the router, only the serialized string is passed to/from the clients.
+ */
+public class PageCursor {
+    private static final int NOT_SET = Integer.MIN_VALUE;
+    private static final long TIME_0 = ZonedDateTime.of(2020, 1, 1, 0, 0, 0, 0, ZoneId.of("UTC")).toEpochSecond();
+    private static final Logger LOG = LoggerFactory.getLogger(PageCursor.class);
+
+    public final Instant earliestDepartureTime;
+    public final Instant latestArrivalTime;
+    public final Duration searchWindow;
+    public final boolean reverseFilteringDirection;
+
+    private PageCursor(
+            Instant earliestDepartureTime,
+            Instant latestArrivalTime,
+            Duration searchWindow,
+            boolean reverseFilteringDirection
+    ) {
+        this.earliestDepartureTime = earliestDepartureTime;
+        this.latestArrivalTime = latestArrivalTime;
+        this.searchWindow = searchWindow;
+        this.reverseFilteringDirection = reverseFilteringDirection;
+    }
+
+    public static PageCursor arriveByCursor(
+        Instant earliestDepartureTime,
+        Instant latestArrivalTime,
+        Duration searchWindow,
+        boolean reverseFilteringDirection
+    ) {
+        return new PageCursor(
+            earliestDepartureTime,
+            latestArrivalTime,
+            searchWindow,
+            reverseFilteringDirection
+        );
+    }
+
+    public static PageCursor departAfterCursor(
+        Instant earliestDepartureTime,
+        Duration searchWindow,
+        boolean reverseFilteringDirection
+    ) {
+        return new PageCursor(
+            earliestDepartureTime,
+            null,
+            searchWindow,
+            reverseFilteringDirection
+        );
+    }
+
+    @Override
+    public String toString() {
+        return ToStringBuilder.of(PageCursor.class)
+                .addTime("edt", earliestDepartureTime)
+                .addTime("lat", latestArrivalTime)
+                .addDuration("searchWindow", searchWindow)
+                .addBoolIfTrue("reverseFilteringDirection", reverseFilteringDirection)
+                .toString();
+    }
+
+    @Nullable
+    public String encode() {
+        var buf = new ByteArrayOutputStream();
+        try(var out = new ObjectOutputStream(buf)){
+            // The order must be the same in the encode and decode function
+            writeTime(earliestDepartureTime, out);
+            writeTime(latestArrivalTime, out);
+            writeDuration(searchWindow, out);
+            writeBoolean(reverseFilteringDirection, out);
+            out.flush();
+            return Base64.getUrlEncoder().encodeToString(buf.toByteArray());
+        }
+        catch (IOException e) {
+            LOG.error("Failed to encode page cursor", e);
+            return null;
+        }
+    }
+
+    @Nullable
+    public static PageCursor decode(String cursor) {
+        if(cursor == null) { return null; }
+
+        var buf = Base64.getUrlDecoder().decode(cursor);
+        var input = new ByteArrayInputStream(buf);
+
+        try(var in = new ObjectInputStream(input)) {
+            // The order must be the same in the encode and decode function
+            var edt = readTime(in);
+            var lat = readTime(in);
+            var searchWindow = readDuration(in);
+            var reverseFilteringDirection = readBoolean(in);
+            return new PageCursor(edt, lat, searchWindow, reverseFilteringDirection);
+        }
+        catch (IOException e) {
+            LOG.error("Unable to decode page cursor: '" + cursor + "'", e);
+            return null;
+        }
+    }
+
+    private Instant getEarliestDepartureTime() {
+        return earliestDepartureTime;
+    }
+
+    public Instant getLatestArrivalTime() {
+        return latestArrivalTime;
+    }
+
+    private static void writeTime(Instant time, ObjectOutputStream out) throws IOException {
+        out.writeInt(time == null ? NOT_SET : (int)(time.getEpochSecond() - TIME_0));
+    }
+
+    @Nullable
+    private static Instant readTime(ObjectInputStream in) throws IOException {
+        var value = in.readInt();
+        return value == NOT_SET ? null : Instant.ofEpochSecond(TIME_0 + value);
+    }
+
+    private static void writeDuration(Duration duration, ObjectOutputStream out) throws IOException {
+        out.writeInt((int)duration.toSeconds());
+    }
+
+    private static Duration readDuration(ObjectInputStream in) throws IOException {
+        return Duration.ofSeconds(in.readInt());
+    }
+
+    private static void writeBoolean(boolean value, ObjectOutputStream out) throws IOException {
+        out.writeBoolean(value);
+    }
+
+    private static boolean readBoolean(ObjectInputStream in) throws IOException {
+        return in.readBoolean();
+    }
+
+    private boolean arriveBy() {
+        return latestArrivalTime != null;
+    }
+}