+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/router/street/DirectFlexRouter.java
@@ -1,15 +1,14 @@
 package org.opentripplanner.routing.algorithm.raptor.router.street;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.List;
 import org.opentripplanner.ext.flex.FlexRouter;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.routing.graphfinder.NearbyStop;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
 import org.opentripplanner.standalone.server.Router;
 
 public class DirectFlexRouter {
@@ -40,7 +39,7 @@ public class DirectFlexRouter {
       FlexRouter flexRouter = new FlexRouter(
               router.graph,
               router.routerConfig.flexParameters(request),
-              directRequest.getDateTime().toInstant(),
+              directRequest.getDateTimeCurrentPage(),
               directRequest.arriveBy,
               directRequest.additionalSearchDaysBeforeToday,
               directRequest.additionalSearchDaysAfterToday,