+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/router/street/FlexAccessEgressRouter.java
@@ -1,14 +1,13 @@
 package org.opentripplanner.routing.algorithm.raptor.router.street;
 
+import java.util.Collection;
+import java.util.List;
 import org.opentripplanner.ext.flex.FlexAccessEgress;
+import org.opentripplanner.ext.flex.FlexParameters;
 import org.opentripplanner.ext.flex.FlexRouter;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.routing.graphfinder.NearbyStop;
-
-import java.util.Collection;
-import java.util.List;
-import org.opentripplanner.ext.flex.FlexParameters;
 
 public class FlexAccessEgressRouter {
 
@@ -35,7 +34,7 @@ public class FlexAccessEgressRouter {
     FlexRouter flexRouter = new FlexRouter(
         request.rctx.graph,
         params,
-        request.getDateTime().toInstant(),
+        request.getDateTimeCurrentPage(),
         request.arriveBy,
         request.additionalSearchDaysBeforeToday,
         request.additionalSearchDaysAfterToday,