+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/TripPatternWithRaptorStopIndexes.java
@@ -44,6 +44,10 @@ public class TripPatternWithRaptorStopIndexes {
         return pattern.getMode();
     }
 
+    public String getNetexSubmode() {
+        return pattern.getNetexSubmode();
+    }
+
     public int[] getStopIndexes() {
         return stopIndexes;
     }