+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/request/RaptorRoutingRequestTransitData.java
@@ -1,6 +1,5 @@
 package org.opentripplanner.routing.algorithm.raptor.transit.request;
 
-import java.time.Instant;
 import java.time.ZonedDateTime;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
@@ -58,7 +57,7 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
   public RaptorRoutingRequestTransitData(
       TransferService transferService,
       TransitLayer transitLayer,
-      Instant departureTime,
+      ZonedDateTime searchStartTime,
       int additionalPastSearchDays,
       int additionalFutureSearchDays,
       TransitDataProviderFilter filter,
@@ -67,14 +66,14 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
 
     this.transferService = transferService;
     this.transitLayer = transitLayer;
+    this.startOfTime = searchStartTime;
 
     // Delegate to the creator to construct the needed data structures. The code is messy so
     // it is nice to NOT have it in the class. It isolate this code to only be available at
     // the time of construction
     var transitDataCreator = new RaptorRoutingRequestTransitDataCreator(
-            transitLayer, departureTime
+            transitLayer, searchStartTime
     );
-    this.startOfTime = transitDataCreator.getSearchStartTime();
     this.activeTripPatternsPerStop = transitDataCreator.createTripPatternsPerStop(
         additionalPastSearchDays,
         additionalFutureSearchDays,
@@ -165,7 +164,4 @@ public class RaptorRoutingRequestTransitData implements RaptorTransitDataProvide
     return validTransitDataEndTime;
   }
 
-  public ZonedDateTime getStartOfTime() {
-    return startOfTime;
-  }
 }